close_range: use unsigned int in protected-fd range guard
The kernel's close_range syscall takes (unsigned int first, unsigned
int last) and walks the fd table whenever first <= last as unsigned.
sanitise_close_range's protected-fd truncation passed the bounds
through lowest_protected_fd_in_range() as signed int, so an a2 of
(unsigned long)-1 -- gen_arg_fd's exhaustion fallback -- arrived as
int -1, tripped the helper's hi < lo guard, and returned -1 (no
protected fd in range).  The kernel then saw close_range(a1,
0xFFFFFFFF) and walked the entire fd table, closing the kcov PC and
CMP fds at KCOV_FD_HIGH_BASE.  Every later kcov_pc_enable /
kcov_pc_disable on that child then returned EBADF, silently
under-counting coverage for the rest of the run.

The same signed-int gap hid the close in the diag: the chronicle
check kcov_chronicle_slot_touched_protected() cast s->a1/a2 to int
before calling the helper, so the per-EBADF latch reported
"closer=close_range" without the /prot suffix that should mark a
fuzz-syscall as the cause.

Match the kernel ABI throughout the protected-fd path:

  - lowest_protected_fd_in_range() now takes (unsigned int lo,
    unsigned int hi) and compares the protected-fd values against
    them as unsigned int.  All currently-protected fds (STDERR_FILENO,
    the stderr capture memfd, kcov.fd, kcov.cmp_fd) fit in int, so
    the return type stays int with -1 for "no protected fd in
    range".

  - sanitise_close_range gates the truncation on (unsigned int) a1
    <= (unsigned int) a2 and casts the bounds to unsigned int before
    handing them to the helper, so the kernel-side walk is bounded
    by lowest_protected - 1 (or inverted to EINVAL) regardless of
    whether gen_arg_fd handed back -1.

  - kcov_chronicle_slot_touched_protected() does the same cast on
    the chronicle slot's saved a1/a2, so the diag /prot suffix now
    correctly fires for any close_range whose range covered a
    protected fd.

Gates: CFLAGS="-g -O2 -D_FORTIFY_SOURCE=2 -Werror" make -j16 clean &&
make -j16 clean.  scripts/check-static.sh: 27/27 pass.  ./trinity
--dry-run -N 200 -C2 --no-cgroup: rc=0.  ./trinity --dry-run -N 5000
-C2 --no-cgroup -c close_range -c close: rc=0, no
pc_enable=EBADF/pc_disable=EBADF/close_range_protect diag emitted.

+++ fds/fds.c
@@ -691,7 +691,20 @@ bool fd_is_protected(int fd)
 	return false;
 }
 
-int lowest_protected_fd_in_range(int lo, int hi)
+/*
+ * Bounds typed as unsigned int to mirror the kernel's close_range ABI:
+ * the syscall takes (unsigned int first, unsigned int last) and walks
+ * the fd table when first <= last as unsigned.  A signed int
+ * comparison here treats rec->a2 == (unsigned long)-1 (the gen_arg_fd
+ * exhaustion fallback, which surfaces to the kernel as 0xFFFFFFFF) as
+ * a negative "hi", makes hi < lo true, and returns -1 -- skipping the
+ * truncation that should have stopped the kernel-side walk before it
+ * reached the kcov fd at KCOV_FD_HIGH_BASE.  All real protected fds
+ * fit in int (STDERR_FILENO is 2; KCOV_FD_HIGH_BASE is 60000; the
+ * stderr memfd is well below INT_MAX), so widening the comparison
+ * operand to unsigned int doesn't shrink the range we cover.
+ */
+int lowest_protected_fd_in_range(unsigned int lo, unsigned int hi)
 {
 	struct childdata *child;
 	int memfd;
@@ -700,22 +713,26 @@ int lowest_protected_fd_in_range(int lo, int hi)
 	if (hi < lo)
 		return -1;
 
-	if (STDERR_FILENO >= lo && STDERR_FILENO <= hi)
+	if ((unsigned int) STDERR_FILENO >= lo &&
+	    (unsigned int) STDERR_FILENO <= hi)
 		lowest = STDERR_FILENO;
 
 	memfd = trinity_stderr_memfd();
-	if (memfd >= 0 && memfd >= lo && memfd <= hi)
+	if (memfd >= 0 &&
+	    (unsigned int) memfd >= lo && (unsigned int) memfd <= hi)
 		if (lowest < 0 || memfd < lowest)
 			lowest = memfd;
 
 	child = this_child();
 	if (child != NULL) {
 		if (child->kcov.fd >= 0 &&
-		    child->kcov.fd >= lo && child->kcov.fd <= hi)
+		    (unsigned int) child->kcov.fd >= lo &&
+		    (unsigned int) child->kcov.fd <= hi)
 			if (lowest < 0 || child->kcov.fd < lowest)
 				lowest = child->kcov.fd;
 		if (child->kcov.cmp_fd >= 0 &&
-		    child->kcov.cmp_fd >= lo && child->kcov.cmp_fd <= hi)
+		    (unsigned int) child->kcov.cmp_fd >= lo &&
+		    (unsigned int) child->kcov.cmp_fd <= hi)
 			if (lowest < 0 || child->kcov.cmp_fd < lowest)
 				lowest = child->kcov.cmp_fd;
 	}
